Start spring security

+++ src/main/java/com/vint/shop/repository/RoleRepository.java
@@ -3,5 +3,7 @@ package com.vint.shop.repository;
 import com.vint.shop.domain.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository <Role, Integer> {
+public interface RoleRepository extends JpaRepository <Role, Long> {
+//    Role findRoleByName(String role);
+//    String findRoleById(int id);
 }
